Module flowcore::lib_provider

lib_provider is used to resolve library references of the type “lib://” using lib search path

Structs

MetaProvider
The MetaProvider implements the Provider trait and based on the url and it’s resolution to a real location for content invokes one of the child providers it has to fetch the content (e.g. File or Http).

Traits

LibProvider
A content provider is responsible with interfacing with the environment and doing IO or what is required to supply content related with flows - isolating other libraries from the File SSystem or IO. It must implement the Provider trait

Provider
A content provider is responsible with interfacing with the environment and doing IO or what is required to supply content related with flows - isolating other libraries from the File SSystem or IO. It must implement the Provider trait