Struct flowcore::model::flow_definition::FlowDefinition

pub struct FlowDefinition {

    pub name: Name,
    pub inputs: IOSet,
    pub outputs: IOSet,
    pub process_refs: Vec<ProcessReference>,
    pub connections: Vec<Connection>,
    pub metadata: MetaData,
    pub docs: String,
    pub alias: Name,
    pub id: usize,
    pub source_url: Url,
    pub route: Route,
    pub subprocesses: BTreeMap<Name, Process>,
    pub lib_references: BTreeSet<Url>,
    pub context_references: BTreeSet<Url>,

}

FlowDefinition defines (at compile time) a parent or child flow in the nested flow hierarchy

Fields
name: Name

name given to this flow

inputs: IOSet

inputs that this flow defines

outputs: IOSet

outputs that this flow defines

process_refs: Vec<ProcessReference>

Set of sub-processes referenced (used) in this flow

connections: Vec<Connection>

connections within this flow, from flow input or to flow outputs

metadata: MetaData

metadata about flow author, versions etc

docs: String

Name of any docs file associated with this Flow

alias: Name

When the same process is used multiple times within a single flow, to disambiguate between them each one must be given an alias that is used to refer to it

id: usize

flows are assigned a numeric id in the hierarchy

source_url: Url

source_url is the url of the file/resource where this flow definition was read from

route: Route

route defines the location in the hierarchy of flows where this ones resides

subprocesses: BTreeMap<Name, Process>

subprocesses are the loaded definition of the processes reference (used) within this flow

lib_references: BTreeSet<Url>

lib_references is the set of library references used in this flow

context_references: BTreeSet<Url>

context_references is the set of context functions used in this flow

Implementations

impl FlowDefinition

pub fn default_url() -> Url

Return a default value for a Url as part of a flow

pub fn set_alias(&mut self, alias: &Name)

Set the alias of this flow to the supplied Name

pub fn get_docs(&self) -> &str

Get the name of any associated docs file

pub fn inputs(&self) -> &IOSet

Get a reference to the set of inputs this flow defines

pub fn inputs_mut(&mut self) -> &mut IOSet

Get a mutable reference to the set of inputs this flow defines

pub fn outputs(&self) -> &IOSet

Get a reference to the set of outputs this flow defines

pub fn set_connection_origin_flow_ids(&mut self)

Set the origin id of all connections to be this flow

pub fn config(
&mut self,
source_url: &Url,
parent_route: &Route,
alias_from_reference: &Name,
id: usize,
initializations: &BTreeMap<String, InputInitializer>
) -> Result<()>

Configure a flow with additional information after it is deserialized from file

pub fn is_runnable(&self) -> bool

Check if the flow can be run (it could be a sub-flow not a context level runnable flow)

pub fn build_connections(&mut self, level: usize) -> Result<()>

Iterate over all the connections defined in the flow, and attempt to connect the source and destination (within the flow), checking the two types are compatible