Struct flowcore::model::io::IO

pub struct IO { /* private fields */ }

IO contains information about the Input or Output of a Function or Flow

Implementations

impl IO

pub fn new<R: Into<Route>>(datatypes: Vec<DataType>, route: R) -> Self

Create a new IO with a specific datatype and at a specific Route

pub fn new_named<R: Into<Route>, N: Into<Name>>(
datatypes: Vec<DataType>,
route: R,
name: N
) -> Self

Create a new IO with a specific datatype and at a specific Route and a Name

pub fn flow_io(&self) -> bool

Is this IO an input or an output of a Flow?

pub fn function_io(&self) -> bool

Is this IO an input or an output of a Function?

pub fn io_type(&self) -> &IOType

Return a reference to the IOType of this IO

pub fn set_io_type(&mut self, io_type: IOType)

Set the IO type

pub fn datatypes(&self) -> &Vec<DataType>

Return a reference to the data type this IO generates or accepts

pub fn set_route(&mut self, route: &Route, io_type: &IOType)

Set the route where this IO resides in the flow hierarchy

pub fn set_datatypes(&mut self, datatypes: &[DataType])

Set the datatypes of this IO

pub fn get_initializer(&self) -> &Option<InputInitializer>

Get a reference to the input initializer of this IO

pub fn get_flow_initializer(&self) -> &Option<InputInitializer>

Get a reference to the flow propagated input initializer of this IO

pub fn set_initializer(
&mut self,
function_initializer: Option<InputInitializer>
) -> Result<()>

Set the input initializer of this IO

pub fn set_flow_initializer(
&mut self,
flow_initializer: Option<InputInitializer>
) -> Result<()>

Set the flow input initializer of this IO