Module flowcore::model::route

route defines a location in the hierarchy of a flow and can locate a flow, a function, or one of its inputs or outputs

Structs

Route

A Route is a String that refers to a particular location within the flow hierarchy and can be used to locate a function, flow, input or output uniquely

Enums

RouteType

A Route can be of various Types

Traits

HasRoute

A trait implemented by objects that have Routes

SetIORoutes

structs with IOs will be able to have the IOs routes set by using parent route

SetRoute

Some structs with Routes will be able to have their route set by using parent route