Trait flowcore::model::route::HasRoute

pub trait HasRoute {
    fn route(&self) -> &Route;
    fn route_mut(&mut self) -> &mut Route;
}

A trait implemented by objects that have Routes

Required Methods

fn route(&self) -> &Route

Return a reference to the Route of the struct that implements this trait

fn route_mut(&mut self) -> &mut Route

Return a mutable reference to the Route of the struct that implements this trait