Trait flowcore::model::route::SetRoute

pub trait SetRoute {
    // Required method
    fn set_routes_from_parent(&mut self, parent: &Route);
}

Some structs with Routes will be able to have their route set by using parent route

Required Methods

fn set_routes_from_parent(&mut self, parent: &Route)

Set the routes in fields of this struct based on the route of it’s parent.