Struct flowcore::model::runtime_function::RuntimeFunction

pub struct RuntimeFunction { /* private fields */ }

RuntimeFunction contains all the information needed about a function and its implementation to be able to execute a flow using it.

Implementations

impl RuntimeFunction

pub fn new<N: Into<String>, R: Into<String>, I: Into<String>>(
name: N,
route: R,
implementation_location: I,
inputs: Vec<Input>,
id: usize,
flow_id: usize,
output_connections: &[OutputConnection],
include_destination_routes: bool
) -> Self

Create a new RuntimeFunction with the specified name, route, implementation etc. This only needs to be used by compilers or IDE generating manifests with functions The library flowrlib just deserializes them from the manifest The Vector of outputs: Output sub-path (or “”), destination function id, destination function io number, Optional path of destination

pub fn reset(&mut self)

Reset a Function to initial state. Used by a debugger at run-time to reset a function as part of a whole flow reset to run it again.

pub fn name(&self) -> &str

Accessor for a RuntimeFunction name

pub fn route(&self) -> &str

Accessor for a RuntimeFunction route

pub fn id(&self) -> usize

Accessor for a RuntimeFunction id

pub fn get_flow_id(&self) -> usize

Accessor for a RuntimeFunction flow_id

pub fn init(&mut self)

Initialize the function to be ready to be called during flow execution

pub fn init_inputs(&mut self, first_time: bool, flow_idle: bool)

Initialize Inputs that have InputInitializers on them

pub fn implementation_location(&self) -> &str

Accessor for a RuntimeFunction implementation_location

pub fn send(&mut self, io_number: usize, value: Value) -> bool

Send a value or array of values to the specified input of this function

pub fn get_output_connections(&self) -> &Vec<OutputConnection>

Accessor for a RuntimeFunction output_connections field

pub fn get_implementation_location(&self) -> &str

Get a reference to the implementation_location

pub fn set_implementation_url(&mut self, manifest_url: &Url) -> Result<()>

Set the implementation_location, as an absolute Url relative to the manifest_url

pub fn get_implementation_url(&self) -> &Url

Get a reference to the implementation_url

pub fn input_count(&self, input_number: usize) -> usize

Determine if the RuntimeFunction input number input_number is full or not

pub fn input_set_count(&self) -> usize

Returns how many inputs sets are available across all the RuntimeFunction Inputs NOTE: For Impure functions without inputs (that can always run and produce a value) this will return usize::MAX

pub fn can_run(&self) -> bool

Can this function run? Either because: - it has input sets to allow it to run - it has no inputs and so can always run

pub fn inputs(&self) -> &Vec<Input>

Inspect the values of the inputs of a RuntimeFunction

pub fn input(&self, id: usize) -> Option<&Input>

Inspect the value of the input of a RuntimeFunction.

pub fn take_input_set(&mut self) -> Result<Vec<Value>>

Read the values from the inputs and return them for use in executing the RuntimeFunction