Trait flowcore::model::validation::Validate

pub trait Validate {
    fn validate(&self) -> Result<()>;
}

Many structs in the model implement the Validate method which is used to check the description deserialized from file obeys some additional constraints that cannot be expressed in the struct definition in serde

Required Methods

fn validate(&self) -> Result<()>

Validate that a deserialized model data structure is valid for use