Crate flowcore

flowcore defines some core structs and traits used by other flow libraries and implementations

Modules

function
function defines functions that form part of a flow

input
input defines the struct for inputs to functions in a flow

lib_manifest
lib_manifest defines the structs for specifying a Library’s manifest and methods to load it

lib_provider
lib_provider is used to resolve library references of the type “lib://” using lib search path

manifest
manifest is the struct that specifies the manifest of functions in a flow

output_connection
output_connection defines a struct for a function’s output connection

url_helper
Utility functions related to Urls Help take file/url strings from a command line and convert them into URLs (as Strings) with schemes for use with flowlibc and flowrlib.

Structs

Error
The Error type.

Enums

ErrorKind
The kind of an error.

Constants

DONT_RUN_AGAIN
Use DONT_RUN_AGAIN to indicate that a function should not be executed more times

RUN_AGAIN
Use RUN_AGAIN to indicate that a function can be executed more times

Traits

Implementation
An implementation runs with an array of inputs and returns a value (or null) and a bool indicating if it should be ran again.

ResultExt
Additional methods for Result, for easy interaction with this crate.

Type Definitions

Result
Convenient wrapper around std::Result.

RunAgain
Implementations should return a value of type RunAgain to indicate if it should be executed more times in the future.