Module flowcore::model

defines many of the core data structures used across libraries and binaries model module defines a number of core data structures that are used across the compiler and the runtime and macros.

Modules

connection

connection defines the connection between one process output to another process’s input

datatype

datatype specifies the type of data permitted on a input, output or connection

flow_definition

flow is the definition of an entire flow, including children flows

flow_manifest

flow_manifest is the struct that specifies the manifest of functions in a flow

function_definition

function defines a function in a flow or library

input

input defines the struct for inputs to functions in a flow

io

io is the object used to define a process’s inputs or outputs

lib_manifest

lib_manifest defines the structs for specifying a Library’s manifest and methods to load it

metadata

metadata defined structs for flow meta data

metrics

metrics defines a struct for runtime execution metrics

name

name is used to name various objects in the flow model

output_connection

output_connection defines a struct for a function’s output connection

process

process is a generic definition of a function or a flow so a flow referring to it or using it does not need to know or define how it is implemented

process_reference

process_reference is an object used within a flow to reference a process defined elsewhere

route

route defines a location in the hierarchy of a flow and can locate a flow, a function, or one of its inputs or outputs

runtime_function

Definition of RuntimeFunction structure

submission

submissiondefines a struct for submitting flows for execution

validation

Traits used for the validation of Model structs